moved SubmitDelegationICQ
